02_Challenge_00_01_NumberAnalizer/pyproject.toml
[project]
name = "02-challengefor-00-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "python-dotenv>=1.1.1",
]
